+++ integration/spark/src/test/spark3/java/io/openlineage/spark3/agent/lifecycle/plan/column/ColumnLevelLineageUtilsNonV2CatalogTest.java
@@ -8,6 +8,7 @@ import io.openlineage.client.OpenLineage;
 import io.openlineage.spark.agent.EventEmitter;
 import io.openlineage.spark.api.OpenLineageContext;
 import io.openlineage.spark3.agent.utils.LastQueryExecutionSparkEventListener;
+import java.util.Arrays;
 import java.util.Optional;
 import lombok.SneakyThrows;
 import lombok.extern.slf4j.Slf4j;
@@ -17,15 +18,10 @@ import org.apache.spark.sql.SparkSession;
 import org.apache.spark.sql.SparkSession$;
 import org.apache.spark.sql.catalyst.plans.logical.LogicalPlan;
 import org.apache.spark.sql.execution.QueryExecution;
-import org.apache.spark.sql.types.IntegerType$;
-import org.apache.spark.sql.types.Metadata;
-import org.apache.spark.sql.types.StructField;
-import org.apache.spark.sql.types.StructType;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import scala.collection.immutable.HashMap;
 
 @Slf4j
 public class ColumnLevelLineageUtilsNonV2CatalogTest {
@@ -34,12 +30,12 @@ public class ColumnLevelLineageUtilsNonV2CatalogTest {
   OpenLineageContext context;
   QueryExecution queryExecution = mock(QueryExecution.class);
 
-  StructType schema =
-      new StructType(
-          new StructField[] {
-            new StructField("a", IntegerType$.MODULE$, false, new Metadata(new HashMap<>())),
-            new StructField("b", IntegerType$.MODULE$, false, new Metadata(new HashMap<>()))
-          });
+  OpenLineage openLineage = new OpenLineage(EventEmitter.OPEN_LINEAGE_PRODUCER_URI);
+  OpenLineage.SchemaDatasetFacet schemaDatasetFacet =
+      openLineage.newSchemaDatasetFacet(
+          Arrays.asList(
+              openLineage.newSchemaDatasetFacetFieldsBuilder().name("a").type("int").build(),
+              openLineage.newSchemaDatasetFacetFieldsBuilder().name("b").type("int").build()));
 
   @BeforeAll
   @SneakyThrows
@@ -92,7 +88,7 @@ public class ColumnLevelLineageUtilsNonV2CatalogTest {
     LogicalPlan plan = LastQueryExecutionSparkEventListener.getLastExecutedLogicalPlan().get();
     when(queryExecution.optimizedPlan()).thenReturn(plan);
     OpenLineage.ColumnLineageDatasetFacet facet =
-        ColumnLevelLineageUtils.buildColumnLineageDatasetFacet(context, schema).get();
+        ColumnLevelLineageUtils.buildColumnLineageDatasetFacet(context, schemaDatasetFacet).get();
 
     assertColumnDependsOn(facet, "a", "file", "column_non_v2/t1", "a");
     assertColumnDependsOn(facet, "b", "file", "column_non_v2/t1", "b");
@@ -108,7 +104,7 @@ public class ColumnLevelLineageUtilsNonV2CatalogTest {
     LogicalPlan plan = LastQueryExecutionSparkEventListener.getLastExecutedLogicalPlan().get();
     when(queryExecution.optimizedPlan()).thenReturn(plan);
     OpenLineage.ColumnLineageDatasetFacet facet =
-        ColumnLevelLineageUtils.buildColumnLineageDatasetFacet(context, schema).get();
+        ColumnLevelLineageUtils.buildColumnLineageDatasetFacet(context, schemaDatasetFacet).get();
 
     assertColumnDependsOn(facet, "a", "file", "column_non_v2/t1", "a");
     assertColumnDependsOn(facet, "b", "file", "column_non_v2/t1", "b");